# introduce a micro stdlib for testing