--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -246,6 +246,7 @@
 - [Added a full-blown DSL for builtins][3471]
 - [Lazy evaluation of RHS argument for || and &&][3492]
 - [Drop Core implementation of IR][3512]
+- [Introduce a smaller version of the standard library, just for testing][3531]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -273,6 +274,7 @@
 [3493]: https://github.com/enso-org/enso/pull/3493
 [3505]: https://github.com/enso-org/enso/pull/3505
 [3512]: https://github.com/enso-org/enso/pull/3512
+[3531]: https://github.com/enso-org/enso/pull/3531
 
 # Enso 2.0.0-alpha.18 (2021-10-12)