--- a/engine/polyglot-api/src/main/java/org/enso/polyglot/RuntimeOptions.java
+++ b/engine/polyglot-api/src/main/java/org/enso/polyglot/RuntimeOptions.java
@@ -66,6 +66,11 @@ public class RuntimeOptions {
   private static final OptionDescriptor LANGUAGE_HOME_OVERRIDE_DESCRIPTOR =
       OptionDescriptor.newBuilder(LANGUAGE_HOME_OVERRIDE_KEY, LANGUAGE_HOME_OVERRIDE).build();
 
+  public static final String EDITION_OVERRIDE = optionName("editionOverride");
+  public static final OptionKey<String> EDITION_OVERRIDE_KEY = new OptionKey<>("");
+  private static final OptionDescriptor EDITION_OVERRIDE_DESCRIPTOR =
+      OptionDescriptor.newBuilder(EDITION_OVERRIDE_KEY, EDITION_OVERRIDE).build();
+
   public static final String DISABLE_IR_CACHES = optionName("disableIrCaches");
   public static final OptionKey<Boolean> DISABLE_IR_CACHES_KEY = new OptionKey<>(false);
   private static final OptionDescriptor DISABLE_IR_CACHES_DESCRIPTOR =
@@ -98,6 +103,7 @@ public class RuntimeOptions {
               ENABLE_GLOBAL_SUGGESTIONS_DESCRIPTOR,
               INTERACTIVE_MODE_DESCRIPTOR,
               LANGUAGE_HOME_OVERRIDE_DESCRIPTOR,
+              EDITION_OVERRIDE_DESCRIPTOR,
               INTERPRETER_SEQUENTIAL_COMMAND_EXECUTION_DESCRIPTOR,
               DISABLE_IR_CACHES_DESCRIPTOR,
               WAIT_FOR_PENDING_SERIALIZATION_JOBS_DESCRIPTOR,