--- a/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeErrorsTest.scala
+++ b/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeErrorsTest.scala
@@ -74,7 +74,7 @@ class RuntimeErrorsTest
             .toFile
             .getAbsolutePath
         )
-        .options(RuntimeOptions.EDITION_OVERRIDE, "0.0.0-dev")
+        .option(RuntimeOptions.EDITION_OVERRIDE, "0.0.0-dev")
         .out(out)
         .serverTransport(runtimeServerEmulator.makeServerTransport)
         .build()