--- a/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeProjectContextTest.scala
+++ b/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeProjectContextTest.scala
@@ -22,7 +22,7 @@ class RuntimeProjectContextTest extends AnyWordSpec with Matchers {
           )
           .option(
             RuntimeOptions.LANGUAGE_HOME_OVERRIDE,
-            Paths.get("../../distribution/component").toFile.getAbsolutePath
+            Paths.get("../../test/micro-distribution/component").toFile.getAbsolutePath
           )
           .option(RuntimeOptions.LOG_LEVEL, "WARNING")
           .build()