--- a/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeProjectContextTest.scala
+++ b/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeProjectContextTest.scala
@@ -27,6 +27,7 @@ class RuntimeProjectContextTest extends AnyWordSpec with Matchers {
               .toFile
               .getAbsolutePath
           )
+          .options(RuntimeOptions.EDITION_OVERRIDE, "0.0.0-dev")
           .option(RuntimeOptions.LOG_LEVEL, "WARNING")
           .build()
         context.initialize(LanguageInfo.ID)