--- a/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeServerTest.scala
+++ b/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeServerTest.scala
@@ -72,7 +72,7 @@ class RuntimeServerTest
             .toFile
             .getAbsolutePath
         )
-        .options(RuntimeOptions.EDITION_OVERRIDE, "0.0.0-dev")
+        .option(RuntimeOptions.EDITION_OVERRIDE, "0.0.0-dev")
         .logHandler(logOut)
         .out(out)
         .serverTransport(runtimeServerEmulator.makeServerTransport)