--- a/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeSuggestionUpdatesTest.scala
+++ b/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeSuggestionUpdatesTest.scala
@@ -55,7 +55,7 @@ class RuntimeSuggestionUpdatesTest
         .option(RuntimeOptions.INTERACTIVE_MODE, "true")
         .option(
           RuntimeOptions.LANGUAGE_HOME_OVERRIDE,
-          Paths.get("../../distribution/component").toFile.getAbsolutePath
+          Paths.get("../../test/micro-distribution/component").toFile.getAbsolutePath
         )
         .out(out)
         .serverTransport(runtimeServerEmulator.makeServerTransport)