--- a/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeSuggestionUpdatesTest.scala
+++ b/engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeSuggestionUpdatesTest.scala
@@ -60,6 +60,7 @@ class RuntimeSuggestionUpdatesTest
             .toFile
             .getAbsolutePath
         )
+        .options(RuntimeOptions.EDITION_OVERRIDE, "0.0.0-dev")
         .out(out)
         .serverTransport(runtimeServerEmulator.makeServerTransport)
         .build()