--- a/engine/runtime/src/main/java/org/enso/interpreter/OptionsHelper.java
+++ b/engine/runtime/src/main/java/org/enso/interpreter/OptionsHelper.java
@@ -34,4 +34,12 @@ public class OptionsHelper {
       return Optional.of(option);
     }
   }
+  public static Optional<String> getEditionOverride(TruffleLanguage.Env env) {
+    String option = env.getOptions().get(RuntimeOptions.EDITION_OVERRIDE_KEY);
+    if (option.equals("")) {
+      return Optional.empty();
+    } else {
+      return Optional.of(option);
+    }
+  }
 }