--- a/engine/runtime/src/main/java/org/enso/interpreter/OptionsHelper.java
+++ b/engine/runtime/src/main/java/org/enso/interpreter/OptionsHelper.java
@@ -34,6 +34,7 @@ public class OptionsHelper {
       return Optional.of(option);
     }
   }
+
   public static Optional<String> getEditionOverride(TruffleLanguage.Env env) {
     String option = env.getOptions().get(RuntimeOptions.EDITION_OVERRIDE_KEY);
     if (option.equals("")) {