--- a/engine/runtime/src/test/scala/org/enso/interpreter/test/InterpreterTest.scala
+++ b/engine/runtime/src/test/scala/org/enso/interpreter/test/InterpreterTest.scala
@@ -109,7 +109,7 @@ class InterpreterContext(
       .in(in)
       .option(
         RuntimeOptions.LANGUAGE_HOME_OVERRIDE,
-        Paths.get("../../distribution/component").toFile.getAbsolutePath
+        Paths.get("../../test/micro-distribution/component").toFile.getAbsolutePath
       )
       .serverTransport { (uri, peer) =>
         if (uri.toString == DebugServerInfo.URI) {