--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/manifest.yaml
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/manifest.yaml
@@ -1,0 +1,3 @@
+archives:
+- main.tgz
+dependencies: []