--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/package.yaml
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/package.yaml
@@ -1,0 +1,10 @@
+name: Base
+namespace: Standard
+version: 0.0.0-dev
+license: APLv2
+authors:
+  - name: Enso Team
+    email: contact@enso.org
+maintainers:
+  - name: Enso Team
+    email: contact@enso.org