--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Any.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Any.enso
@@ -1,0 +1,3 @@
+type Any
+    @Builtin_Type
+    type Any