--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Boolean.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Boolean.enso
@@ -1,0 +1,11 @@
+type Boolean
+    @Builtin_Type
+    type Boolean
+
+    if_then_else ~on_true ~on_false = @Builtin_Method "Boolean.if_then_else"
+
+@Builtin_Type
+type True
+
+@Builtin_Type
+type False