--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/List.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/List.enso
@@ -1,0 +1,3 @@
+type List
+    type Nil
+    type Cons x xs