--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Numbers.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Numbers.enso
@@ -1,0 +1,7 @@
+type Number
+    @Builtin_Type
+    type Number
+
+type Integer
+    @Builtin_Type
+    type Integer