--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Text.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Data/Text.enso
@@ -1,0 +1,3 @@
+type Text
+    @Builtin_Type
+    type Text