--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/IO.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/IO.enso
@@ -1,0 +1,3 @@
+print_err message = @Builtin_Method "IO.print_err"
+println message = @Builtin_Method "IO.println"
+readln = @Builtin_Method "IO.readln"