--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Main.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Main.enso
@@ -1,0 +1,20 @@
+import project.IO
+export project.IO
+
+import project.Data.List
+from project.Data.List export Nil, Cons, List
+
+import project.Polyglot
+from project.Polyglot export all
+
+export project.Polyglot.Java
+import project.Polyglot.Java
+
+import project.Data.Array
+from project.Data.Array export Array
+
+import project.Error.Common
+from project.Error.Common export all
+
+import project.Data.Numbers
+from project.Data.Numbers export Number