--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Nothing.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Nothing.enso
@@ -1,0 +1,3 @@
+type Nothing
+    @Builtin_Type
+    type Nothing