--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Polyglot.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Polyglot.enso
@@ -1,0 +1,13 @@
+type Polyglot
+    @Builtin_Type
+    type Polyglot
+get_array_size array = @Builtin_Method "Polyglot.get_array_size"
+execute callable arguments = @Builtin_Method "Polyglot.execute"
+get_member object member_name = @Builtin_Method "Polyglot.get_member"
+get_members object = @Builtin_Method "Polyglot.get_members"
+new constructor arguments = @Builtin_Method "Polyglot.new"
+invoke target name arguments = @Builtin_Method "Polyglot.invoke"
+has_source_location value = @Builtin_Method "Polyglot.has_source_location"
+get_source_location value = @Builtin_Method "Polyglot.get_source_location"
+is_language_installed language_name = @Builtin_Method "Polyglot.is_language_installed"
+get_executable_name = @Builtin_Method "Polyglot.get_executable_name"