--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Polyglot/Java.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Polyglot/Java.enso
@@ -1,0 +1,1 @@
+lookup_class name = @Builtin_Method "Java.lookup_class"