--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime/Debug.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime/Debug.enso
@@ -1,0 +1,2 @@
+eval expression = @Builtin_Method "Debug.eval"
+breakpoint = @Builtin_Method "Debug.breakpoint"