--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime/Resource.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime/Resource.enso
@@ -1,0 +1,10 @@
+bracket : Any -> (Any -> Nothing) -> (Any -> Any) -> Any
+bracket ~constructor ~destructor ~action = @Builtin_Method "Resource.bracket"
+
+type Managed_Resource
+    @Builtin_Type
+    type Managed_Resource
+    register resource function = @Builtin_Method "Managed_Resource.register"
+    finalize = @Builtin_Method "Managed_Resource.finalize"
+    with ~action = @Builtin_Method "Managed_Resource.with"
+    take = @Builtin_Method "Managed_Resource.take"