--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime/Thread.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/Runtime/Thread.enso
@@ -1,0 +1,1 @@
+with_interrupt_handler ~action ~interrupt_handler = @Builtin_Method "Thread.with_interrupt_handler"