--- a/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/System.enso
+++ b/test/micro-distribution/lib/Standard/Base/0.0.0-dev/src/System.enso
@@ -1,0 +1,4 @@
+create_process command arguments input redirect_in redirect_out redirect_err = @Builtin_Method "System.create_process"
+
+@Builtin_Type
+type System_Process_Result exit_code stdout stderr