@JaroslavTulach 2022-06-16
Clever! Reminds me of my efforts to define [minimal Java](http://hudson.apidesign.org/hudson/job/bck2brwsr.javadoc/lastSuccessfulBuild/artifact/rt/emul/mini/target/site/apidocs/index.html).

Once I suggested that `runtime.jar` shall carry a _minimalistic_ library inside of itself to be usable as a standalone JAR. If we packaged this micro stdlib into the `runtime.jar`, I would have what I wanted!